Seeders y fakers parte 1

+++ database/migrations/2018_11_28_212712_create_hotels_table.php
@@ -18,7 +18,7 @@ class CreateHotelsTable extends Migration
             $table->string('nombre',45);
             $table->string('telefono',20);
             $table->string('compania',70);
-            $table->decimal('calificacion');
+            $table->integer('calificacion');
             $table->text('descripcion');
             $table->timestamps();
